Step and Optimize Logic
